# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para plataforma de serviços de limpeza.",
        "title": "Phresh-Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/limpezas/{limpeza_id}/solicitacoes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Listar solicitações de uma limpeza (faxineiro)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "limpeza_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaSolicitacao"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/solicitacoes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Criar solicitação de limpeza",
                "parameters": [
                    {
                        "description": "Dados da solicitação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequisicaoCriarSolicitacao"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaSolicitacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/solicitacoes/{cliente_id}/{limpeza_id}/aceitar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Aceitar solicitação de limpeza",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do cliente",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "limpeza_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaSolicitacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/solicitacoes/{cliente_id}/{limpeza_id}/rejeitar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Rejeitar solicitação de limpeza",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do cliente",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "limpeza_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaSolicitacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
        "/solicitacoes/{limpeza_id}/cancelar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Cancelar solicitação de limpeza",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da limpeza",
                        "name": "limpeza_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaSolicitacao"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
        "/usuarios/eu/solicitacoes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "solicitacoes"
                ],
                "summary": "Listar solicitações do cliente autenticado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RespostaSolicitacao"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RespostaErro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.EnderecoDTO": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cidade": {
                    "type": "string"
                },
                "complemento": {
                    "type": "string"
                },
                "estado": {
                    "type": "string"
                },
                "rua": {
                    "type": "string"
                }
            }
        },
...
        "dto.RequisicaoCriarSolicitacao": {
            "type": "object",
            "properties": {
                "data_agendada": {
                    "type": "string"
                },
                "limpeza_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.RespostaErro": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "detalhes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mensagem": {
                    "type": "string"
                }
            }
        },
...
        "dto.RespostaSolicitacao": {
            "type": "object",
            "properties": {
                "cliente_id": {
                    "type": "integer"
                },
                "data_agendada": {
                    "type": "string"
                },
                "endereco": {
                    "$ref": "#/definitions/dto.EnderecoDTO"
                },
                "id": {
                    "type": "integer"
                },
                "limpeza_id": {
                    "type": "integer"
                },
                "multa_cancelamento": {
                    "type": "number"
                },
                "preco_total": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}